Add files via upload
This project was to build a simple object tracker using a
Detectron2 pre-trained model by performing an instance segmentation task on a set of frames of
a video in order to obtain the predictions. Then, we implemented an algorithm to link these
predictions over time to perform an object tracking task.